ref: moved types to their own dedicated types files
