+++ app/favorites/page.tsx
@@ -5,22 +5,8 @@ import axios, { type AxiosResponse } from "axios";
 import Header from "@/components/Header";
 import MovieCard from "@/components/MovieCard";
 import { useToast } from "@/components/ToastProvider";
-import { MovieDetail } from "../movie/[id]/page";
+import { MovieDetail } from "../movie/[id]/types";
 import { fetchFavorites, fetchWatchLater } from "@/lib/pageFetches";
-
-export interface FavoriteMovie {
-  id: string;
-  movieId: number;
-  createdAt: string;
-  movie: {
-    id: number;
-    title: string;
-    overview: string;
-    poster_path: string;
-    vote_average: number;
-    release_date: string;
-  };
-}
 
 export default function FavoritesPage() {
   const [favorites, setFavorites] = useState<MovieDetail[]>([]);
