+++ app/movie/[id]/page.tsx
@@ -6,26 +6,7 @@ import axios from "axios";
 import Header from "@/components/Header";
 import VideoPlayer from "@/components/VideoPlayer";
 import { useToast } from "@/components/ToastProvider";
-
-export interface MovieDetail {
-  id: number;
-  title: string;
-  overview: string;
-  poster_path: string;
-  backdrop_path: string;
-  vote_average: number;
-  release_date: string;
-  runtime: number;
-  genres: Array<{ id: number; name: string }>;
-  videos: {
-    results: Array<{
-      key: string;
-      name: string;
-      site: string;
-      type: string;
-    }>;
-  };
-}
+import { MovieDetail } from "./types";
 
 export default function MovieDetailPage() {
   const params = useParams();
